requirements.txt
fastapi
uvicorn[standard]
google-genai
supabase
python-dotenv
pydantic
pypdf
